services/stt/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
faster-whisper==1.0.1
numpy==1.26.3
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1

# Error correction and clarification
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
loguru>=0.7.2
